pdt/requirements.txt
Django==5.1.4
psycopg[binary]==3.2.3
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
redis==5.2.1
django-allauth==65.3.1
django-environ==0.12.0
django-htmx==1.21.0
whitenoise==6.8.2
Pillow==11.1.0
gunicorn==23.0.0
